Refine R-D optimization of coeffs
STATS_CHANGED

+++ av1/encoder/encoder.h
@@ -960,7 +960,7 @@ typedef struct AV1_COMP {
   int max_comp_type_rd_threshold_mul;
   int max_comp_type_rd_threshold_div;
 
-  unsigned int tx_domain_dist_threshold;
+  double tx_domain_dist_threshold;
 
   // Factor to control R-D optimization of coeffs based on block
   // mse.
